scraper/requirements.txt
selenium==4.38.0
webdriver-manager==4.0.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-dateutil==2.9.0
supabase==2.23.0